MONGOID-4427 Don't unset an embeds_one document because it will be replaced by the new document

+++ lib/mongoid/persistable/deletable.rb
@@ -21,10 +21,12 @@ module Mongoid
       def delete(options = {})
         raise Errors::ReadonlyDocument.new(self.class) if readonly?
         prepare_delete do
-          if embedded?
-            delete_as_embedded(options)
-          else
-            delete_as_root
+          unless options[:persist] == false
+            if embedded?
+              delete_as_embedded(options)
+            else
+              delete_as_root
+            end
           end
         end
       end
